requirements.txt
# Slack
slack-bolt>=1.27.0
slack-sdk>=3.40.1

# Environment
python-dotenv>=1.2.0
certifi>=2022.9.24

# RAG / Vector Store
chromadb>=0.4.0

# HTTP
requests>=2.28.0

# SSH tunnel (LLM connect script)
paramiko>=4.0.0
bcrypt>=5.0.0
cryptography>=46.0.5
PyNaCl>=1.6.2
pycparser>=3.0
cffi>=2.0.0
invoke>=2.2.0
